Avoid signed/unsigned comparison warning
Summary:
!ci_branch_mk2
Ref T72249

If customers compile their custom ops with -Wall then this warning appears which seems undesirable.
Cast the unsigned vector size to a signed so it can be compared against the signed i.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: shirazb

Maniphest Tasks: T72249

Differential Revision: https://phabricator.sourcevertex.net/D79795

+++ willow/include/popart/tensorinfo.hpp
@@ -212,7 +212,7 @@ public:
   int64_t nbytes() const;
   // Defined in-header to encourage inlining (it's called a lot).
   int64_t dim(int i) const {
-    if (i >= shape_v.size()) {
+    if (i >= static_cast<int>(shape_v.size())) {
       throw error(
           "Invalid input dimension {}, tensor of rank {}", i, shape_v.size());
     }
